add pipeline status and remove the old status checker
